+++ src/dashboard/run_demo.py
@@ -170,7 +170,7 @@ def get_user_lookup() -> pd.DataFrame:
         the wallet address itself if a wallet is found without a registered username.
     """
     # Get data
-    agents = postgres.get_agents(session)
+    agents = src.data.hyperdrive.postgres.get_agents(session)
     user_map = postgres.get_user_map(session)
     # Usernames in postgres are bots
     user_map["username"] = user_map["username"] + " (bots)"
@@ -240,7 +240,7 @@ while True:
     txn_data = postgres.get_transactions(session, -max_live_blocks)
     pool_info_data = src.data.hyperdrive.postgres.get_pool_info(session, -max_live_blocks)
     combined_data = get_combined_data(txn_data, pool_info_data)
-    wallet_deltas = postgres.get_wallet_deltas(session)
+    wallet_deltas = src.data.hyperdrive.postgres.get_wallet_deltas(session)
     ticker = get_ticker(wallet_deltas, txn_data, pool_info_data, user_lookup)
 
     (fixed_rate_x, fixed_rate_y) = calc_fixed_rate(combined_data, config_data)
