+++ src/data/postgres.py
@@ -3,10 +3,9 @@
 from __future__ import annotations
 
 import os
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import Type
 
-import numpy as np
 import pandas as pd
 import sqlalchemy
 from sqlalchemy import URL, Engine, MetaData, Table, create_engine, exc, func, inspect
@@ -14,7 +13,6 @@ from sqlalchemy.orm import Session, sessionmaker
 
 from src.data.db_schema import Base, Transaction, UserMap
 from src.data.hyperdrive.db_schema import CheckpointInfo, PoolInfo, WalletDelta, WalletInfo
-from src.data.hyperdrive.postgres import get_pool_info
 
 # classes for sqlalchemy that define table schemas have no methods.
 # pylint: disable=too-few-public-methods
@@ -270,216 +268,6 @@ def get_transactions(session: Session, start_block: int | None = None, end_block
     return pd.read_sql(query.statement, con=session.connection()).set_index("blockNumber")
 
 
-def get_all_wallet_info(session: Session, start_block: int | None = None, end_block: int | None = None) -> pd.DataFrame:
-    """Get all of the wallet_info data in history and returns as a pandas dataframe.
-
-    Arguments
-    ---------
-    session : Session
-        The initialized session object
-    start_block : int | None, optional
-        The starting block to filter the query on. start_block integers
-        matches python slicing notation, e.g., list[:3], list[:-3]
-    end_block : int | None, optional
-        The ending block to filter the query on. end_block integers
-        matches python slicing notation, e.g., list[:3], list[:-3]
-
-    Returns
-    -------
-    DataFrame
-        A DataFrame that consists of the queried wallet info data
-    """
-    query = session.query(WalletInfo)
-
-    # Support for negative indices
-    if (start_block is not None) and (start_block < 0):
-        start_block = get_latest_block_number_from_table(WalletInfo, session) + start_block + 1
-    if (end_block is not None) and (end_block < 0):
-        end_block = get_latest_block_number_from_table(WalletInfo, session) + end_block + 1
-
-    if start_block is not None:
-        query = query.filter(WalletInfo.blockNumber >= start_block)
-    if end_block is not None:
-        query = query.filter(WalletInfo.blockNumber < end_block)
-
-    return pd.read_sql(query.statement, con=session.connection())
-
-
-def get_wallet_info_history(session: Session) -> dict[str, pd.DataFrame]:
-    """Get the history of all wallet info over block time.
-
-    Arguments
-    ---------
-    session : Session
-        The initialized session object
-
-    Returns
-    -------
-    dict[str, DataFrame]
-        A dictionary keyed by the wallet address, where the values is a DataFrame
-        where the index is the block number, and the columns is the number of each
-        token the address has at that block number, plus a timestamp and the share price of the block
-    """
-    # Get data
-    all_wallet_info = get_all_wallet_info(session)
-    pool_info_lookup = get_pool_info(session)[["timestamp", "sharePrice"]]
-
-    # Pivot tokenType to columns, keeping walletAddress and blockNumber
-    all_wallet_info = all_wallet_info.pivot(
-        values="tokenValue", index=["walletAddress", "blockNumber"], columns=["tokenType"]
-    )
-    # Forward fill nans here, as no data means no change
-    all_wallet_info = all_wallet_info.fillna(method="ffill")
-
-    # Convert walletAddress to outer dictionary
-    wallet_info_dict = {}
-    for addr in all_wallet_info.index.get_level_values(0).unique():
-        addr_wallet_info = all_wallet_info.loc[addr]
-        # Reindex block number to be continuous, filling values with the last entry
-        addr_wallet_info = addr_wallet_info.reindex(pool_info_lookup.index, method="ffill")
-        addr_wallet_info["timestamp"] = pool_info_lookup.loc[addr_wallet_info.index, "timestamp"]
-        addr_wallet_info["sharePrice"] = pool_info_lookup.loc[addr_wallet_info.index, "sharePrice"]
-        # Drop all rows where BASE tokens are nan
-        addr_wallet_info = addr_wallet_info.dropna(subset="BASE")
-        # Fill the rest with 0 values
-        addr_wallet_info = addr_wallet_info.fillna(0)
-        # Remove name from column index
-        addr_wallet_info.columns.name = None
-        wallet_info_dict[addr] = addr_wallet_info
-
-    return wallet_info_dict
-
-
-def get_current_wallet_info(
-    session: Session, start_block: int | None = None, end_block: int | None = None
-) -> pd.DataFrame:
-    """Get the balance of a wallet and a given end_block.
-
-    Note
-    ----
-    Here, you can specify a start_block for performance reasons,
-    but if a trade happens before the start_block,
-    that token won't show up in the result.
-
-    Arguments
-    ---------
-    session : Session
-        The initialized session object
-    start_block : int | None, optional
-        The starting block to filter the query on. start_block integers
-        matches python slicing notation, e.g., list[:3], list[:-3]
-    end_block : int | None, optional
-        The ending block to filter the query on. end_block integers
-        matches python slicing notation, e.g., list[:3], list[:-3]
-
-    Returns
-    -------
-    DataFrame
-        A DataFrame that consists of the queried wallet info data
-    """
-    all_wallet_info = get_all_wallet_info(session, start_block=start_block, end_block=end_block)
-    # Get last entry in the table of each wallet address and token type
-    # This should always return a dataframe
-    # Pandas doesn't play nice with types
-    result = (
-        all_wallet_info.sort_values("blockNumber", ascending=False)
-        .groupby(["walletAddress", "tokenType"])
-        .agg(
-            {
-                "tokenValue": "first",
-                "baseTokenType": "first",
-                "maturityTime": "first",
-                "blockNumber": "first",
-                "sharePrice": "first",
-            }
-        )
-    )
-    assert isinstance(result, pd.DataFrame), "result is not a dataframe"
-    current_wallet_info: pd.DataFrame = result
-
-    # Rename blockNumber column
-    current_wallet_info = current_wallet_info.rename({"blockNumber": "latestUpdateBlock"}, axis=1)
-    # Filter current_wallet_info to remove 0 balance tokens
-    current_wallet_info = current_wallet_info[current_wallet_info["tokenValue"] > 0]
-
-    return current_wallet_info
-
-
-def get_wallet_deltas(session: Session, start_block: int | None = None, end_block: int | None = None) -> pd.DataFrame:
-    """Get all wallet_delta data in history and returns as a pandas dataframe.
-
-    Arguments
-    ---------
-    session : Session
-        The initialized session object
-    start_block : int | None, optional
-        The starting block to filter the query on. start_block integers
-        matches python slicing notation, e.g., list[:3], list[:-3]
-    end_block : int | None, optional
-        The ending block to filter the query on. end_block integers
-        matches python slicing notation, e.g., list[:3], list[:-3]
-
-    Returns
-    -------
-    DataFrame
-        A DataFrame that consists of the queried wallet info data
-    """
-    query = session.query(WalletDelta)
-
-    # Support for negative indices
-    if (start_block is not None) and (start_block < 0):
-        start_block = get_latest_block_number_from_table(WalletDelta, session) + start_block + 1
-    if (end_block is not None) and (end_block < 0):
-        end_block = get_latest_block_number_from_table(WalletDelta, session) + end_block + 1
-
-    if start_block is not None:
-        query = query.filter(WalletDelta.blockNumber >= start_block)
-    if end_block is not None:
-        query = query.filter(WalletDelta.blockNumber < end_block)
-
-    return pd.read_sql(query.statement, con=session.connection())
-
-
-def get_agents(session: Session, start_block: int | None = None, end_block: int | None = None) -> list[str]:
-    """Get the list of all agents from the WalletInfo table.
-
-    Arguments
-    ---------
-    session : Session
-        The initialized session object
-    start_block : int | None, optional
-        The starting block to filter the query on. start_block integers
-        matches python slicing notation, e.g., list[:3], list[:-3]
-    end_block : int | None, optional
-        The ending block to filter the query on. end_block integers
-        matches python slicing notation, e.g., list[:3], list[:-3]
-
-    Returns
-    -------
-    list[str]
-        A list of agent addresses
-    """
-    query = session.query(WalletInfo.walletAddress)
-    # Support for negative indices
-    if (start_block is not None) and (start_block < 0):
-        start_block = get_latest_block_number_from_table(WalletInfo, session) + start_block + 1
-    if (end_block is not None) and (end_block < 0):
-        end_block = get_latest_block_number_from_table(WalletInfo, session) + end_block + 1
-
-    if start_block is not None:
-        query = query.filter(WalletInfo.blockNumber >= start_block)
-    if end_block is not None:
-        query = query.filter(WalletInfo.blockNumber < end_block)
-
-    if query is None:
-        return []
-    query = query.distinct()
-
-    results = pd.read_sql(query.statement, con=session.connection())
-
-    return results["walletAddress"].to_list()
-
-
 def get_user_map(session: Session, address: str | None = None) -> pd.DataFrame:
     """Get all usermapping and returns as a pandas dataframe.
 
@@ -543,147 +331,3 @@ def get_latest_block_number_from_table(
     if result[0] is None:
         return 0
     return int(result[0])
-
-
-@dataclass
-class AgentPosition:
-    """Details what the agent holds, how it's changed over time, and how much it's worth.
-
-    Notes
-    -----
-    At a high level "position" refers to the entire portfolio of holdings.
-    The portfolio is comprised of multiple positions, built up through multiple trades over time.
-    - At most, the agent can have positions equal to the number of checkpoints (trades within a checkpoint are fungible)
-    - DataFrames are [blocks, positions] in shape, for convenience and vectorization
-    - Series are [blocks] in shape
-
-    Examples
-    --------
-    To create an agent position you only need to pass in the wallet, from `pg.get_wallet_info_history(session)`:
-
-    >>> agent_position = AgentPosition(pg.get_wallet_info_history(session))
-
-    Use the same index across multiple tables:
-    >>> block = 69
-    >>> position = 3
-    >>> position_name = agent_position.positions.columns[position]
-    >>> holding = agent_position.positions.loc[block, position]
-    >>> open_share_price = agent_position.open_share_price.loc[block, position]
-    >>> pnl = agent_position.pnl.loc[block, position]
-    >>> print(f"agent holds {holding} bonds in {position_name} at block {block} worth {pnl}"})
-    agent holds  55.55555556 bonds in LONG-20240715 at block 69 worth 50
-
-    Attributes
-    ----------
-    positions : pd.DataFrame
-        The agent's holding of a single position, in bonds (# of longs or shorts).
-    deltas : pd.DataFrame
-        Change in each position, from the previous block.
-    open_share_price : pd.DataFrame
-        Weighted average open share price of each position
-    pnl : pd.Series
-        Value of the agent's positions.
-    share_price : pd.Series
-        Share price at the time of the current block.
-    timestamp : pd.Series
-        Timestamp of the current block.
-    """
-
-    positions: pd.DataFrame
-    deltas: pd.DataFrame
-    open_share_price: pd.DataFrame
-    share_price: pd.Series
-    timestamp: pd.Series
-    pnl: pd.Series = field(default_factory=pd.Series)
-    share_price: pd.Series = field(default_factory=pd.Series)
-    timestamp: pd.Series = field(default_factory=pd.Series)
-
-    def __init__(self, wallet_history: pd.DataFrame):
-        """Calculate multiple relevant historical breakdowns of an agent's position."""
-        # Prepare PNL Series filled with NaNs, in the shape of [blocks]
-        self.pnl = pd.Series(data=np.nan, index=wallet_history.index)
-
-        # Scrap the wallet history for parts. First we extract the share price and timestamp.
-        self.share_price = wallet_history["sharePrice"]
-        self.timestamp = wallet_history["timestamp"]
-        # Then we keep track of every other column, to extract them into other tables.
-        other_columns = [col for col in wallet_history.columns if col not in ["sharePrice", "timestamp"]]
-
-        # Create positions dataframe which tracks aggregate holdings.
-        self.positions = wallet_history.loc[:, other_columns].copy()
-
-        # Create deltas dataframe which tracks change in holdings.
-        self.deltas = self.positions.diff()
-        # After the diff() call above, the first row of the deltas table will be NaN.
-        # Replace them with the first row of the positions table, effectively capturing a delta from 0.
-        self.deltas.iloc[0] = self.positions.iloc[0]
-
-        # Prepare tables filled with NaNs, in the shape of [blocks, positions]
-        share_price_on_increases = pd.DataFrame(
-            data=np.nan,  # type: ignore
-            index=self.deltas.index,  # type: ignore
-            columns=self.deltas.columns,  # type: ignore
-        )
-        self.open_share_price = pd.DataFrame(
-            data=np.nan,  # type: ignore
-            index=self.deltas.index,  # type: ignore
-            columns=self.deltas.columns,  # type: ignore
-        )
-
-        # When we have an increase in position, we use the current block's share_price
-        share_price_on_increases = share_price_on_increases.mask(self.deltas > 0, self.share_price, axis=0)
-
-        # Fill forward to replace NaNs. Table is now full of share prices, sourced only from increases in position.
-        share_price_on_increases.fillna(method="ffill", inplace=True, axis=0)
-
-        # Calculate weighted average share price across all deltas (couldn't figure out how to do this vector-wise).
-        # vectorised attempt: ap.open_share_price = (share_price_on_increases * deltas).cumsum(axis=0) / positions
-        # First row of weighted average open share price is equal to the share
-        # price on increases since there's nothing to weight.
-        self.open_share_price.iloc[0] = share_price_on_increases.iloc[0]
-
-        # Now we loop across the remaining rows, updated the weighted averages for positions that change.
-        for row in self.deltas.index[1:]:
-            # An update is required for columns which increase in size this row, identified by a positive delta.
-            update_required = self.deltas.loc[row, :].values > 0  # boolean values where we have positive deltas.
-            cols = self.deltas.columns[update_required].to_list()  # associated column names
-
-            new_price = []
-            if len(update_required) > 0:
-                # calculate update, per this general formula:
-                # new_price = (delta_amount * current_price + old_amount * old_price) / (old_amount + delta_amount)
-                new_price = (
-                    share_price_on_increases.loc[row, cols] * self.deltas.loc[row, cols]  # type: ignore
-                    + self.open_share_price.loc[row - 1, cols] * self.positions.loc[row - 1, cols]  # type: ignore
-                ) / (
-                    self.deltas.loc[row, cols] + self.positions.loc[row - 1, cols]  # type: ignore
-                )
-
-            # Keep previous result where an update isn't required, otherwise replace with new_price
-            self.open_share_price.loc[row, :] = self.open_share_price.loc[row - 1, :].where(
-                ~update_required, new_price, axis=0
-            )
-
-
-def get_agent_positions(session: Session, filter_addr: list[str] | None = None) -> dict[str, AgentPosition]:
-    """Create an AgentPosition for each agent in the wallet history.
-
-    Arguments
-    ---------
-    session : Session
-        The initialized session object
-    filter_addr : list[str] | None
-        Only return these addresses. Returns all if None
-
-    Returns
-    -------
-    dict[str, AgentPosition]
-        Returns a dictionary keyed by wallet address, value of an agent's position
-    """
-    if filter_addr is None:
-        return {agent: AgentPosition(wallet) for agent, wallet in get_wallet_info_history(session).items()}
-    return {
-        agent: AgentPosition(wallet)
-        for agent, wallet in get_wallet_info_history(session).items()
-        if agent in filter_addr
-    }
