+++ src/data/tests/test_wallet_deltas.py
@@ -96,7 +96,7 @@ class TestWalletDeltaInterface:
         wallet_delta_3 = WalletDelta(blockNumber=2, transactionHash="a", delta=Decimal("3.3"))
         src.data.hyperdrive.postgres.add_wallet_deltas([wallet_delta_1, wallet_delta_2, wallet_delta_3], session)
 
-        wallet_delta_df = postgres.get_wallet_deltas(session)
+        wallet_delta_df = src.data.hyperdrive.postgres.get_wallet_deltas(session)
         np.testing.assert_array_equal(wallet_delta_df["delta"], np.array([3.1, 3.2, 3.3]))
 
     def test_block_query_wallet_delta(self, session):
@@ -106,17 +106,17 @@ class TestWalletDeltaInterface:
         wallet_delta_3 = WalletDelta(blockNumber=2, transactionHash="a", delta=Decimal("3.3"))
         src.data.hyperdrive.postgres.add_wallet_deltas([wallet_delta_1, wallet_delta_2, wallet_delta_3], session)
 
-        wallet_delta_df = postgres.get_wallet_deltas(session, start_block=1)
+        wallet_delta_df = src.data.hyperdrive.postgres.get_wallet_deltas(session, start_block=1)
         np.testing.assert_array_equal(wallet_delta_df["delta"], np.array([3.2, 3.3]))
 
-        wallet_delta_df = postgres.get_wallet_deltas(session, start_block=-1)
+        wallet_delta_df = src.data.hyperdrive.postgres.get_wallet_deltas(session, start_block=-1)
         np.testing.assert_array_equal(wallet_delta_df["delta"], np.array([3.3]))
 
-        wallet_delta_df = postgres.get_wallet_deltas(session, end_block=1)
+        wallet_delta_df = src.data.hyperdrive.postgres.get_wallet_deltas(session, end_block=1)
         np.testing.assert_array_equal(wallet_delta_df["delta"], np.array([3.1]))
 
-        wallet_delta_df = postgres.get_wallet_deltas(session, end_block=-1)
+        wallet_delta_df = src.data.hyperdrive.postgres.get_wallet_deltas(session, end_block=-1)
         np.testing.assert_array_equal(wallet_delta_df["delta"], np.array([3.1, 3.2]))
 
-        wallet_delta_df = postgres.get_wallet_deltas(session, start_block=1, end_block=-1)
+        wallet_delta_df = src.data.hyperdrive.postgres.get_wallet_deltas(session, start_block=1, end_block=-1)
         np.testing.assert_array_equal(wallet_delta_df["delta"], np.array([3.2]))
